fixed wine_utils dxup_setup.in; now links to DXVK installation
the setup script assumed that DXVK's d3d11.dll and dxgi.dll (as dxgi_original)
would be in the dxup dll directory. They are not, unless the user puts them there.
So, I added a conditional to link to DXVK's dlls if not already done.
Note: this checks if symlinks exist; so manual redirects have to be symlinks

Also, fixed a typo